undo rename of user_page
