+++ lib/user/user.dart
@@ -1,2 +1,2 @@
 export 'bloc/user_bloc.dart';
-export 'pages/user_page_old.dart';
+export 'pages/user_page.dart';
